+++ projects/ids-enterprise-ng/src/lib/locale/soho-translate.spec.ts
@@ -2,7 +2,7 @@
 
 import { SohoTranslatePipe } from './soho-translate.pipe';
 
-describe('Pipe: SohoTranslatePipe', () => {
+fdescribe('Pipe: SohoTranslatePipe', () => {
   let pipe: SohoTranslatePipe;
 
   beforeEach(() => {
@@ -19,5 +19,24 @@ describe('Pipe: SohoTranslatePipe', () => {
 
   it('providing missing resourceKey returns [resourceKey]', () => {
     expect(pipe.transform('NotAKnownValue')).toBe('[NotAKnownValue]');
+  });
+
+  it('be possible to extend a language', (done) => {
+    Soho.Locale.culturesPath = '/assets/ids-enterprise/js/cultures/';
+    Soho.Locale.set('it-IT')
+    .done(() => {
+      const lang = Soho.Locale.currentLanguage.name;
+      const newStrings = {
+        Thanks: { id: 'Thanks', value: 'Grazie', comment: ''},
+        YourWelcome: { id: 'YourWelcome', value: 'Prego', comment: ''}
+      };
+
+      if (Soho.Locale.languages[lang].messages !== undefined) {
+        Soho.Locale.extendTranslations(lang, newStrings);
+      }
+
+      expect(pipe.transform('Thanks')).toBe('Grazie');
+      done();
+    });
   });
 });